implemented tests (and fixed db creation bugs) for documentAPI and phraseAPI
