+++ src/test/java/care/smith/top/backend/TopBackendApplicationTests.java
@@ -6,7 +6,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ContextConfiguration;
 
 @SpringBootTest
-@ContextConfiguration(initializers = TopBackendContextInitializer.class)
 class TopBackendApplicationTests {
 
   @Test
