+++ src/test/java/care/smith/top/backend/api/nlp/DocumentApiDelegateImplTest.java
@@ -1,32 +1,24 @@
 package care.smith.top.backend.api.nlp;
 
 import care.smith.top.backend.AbstractNLPTest;
-import care.smith.top.backend.model.neo4j.DocumentNodeEntity;
-import care.smith.top.backend.repository.neo4j.DocumentNodeRepository;
-import care.smith.top.backend.service.nlp.DocumentService;
 import care.smith.top.model.Document;
 import care.smith.top.model.DocumentGatheringMode;
 import care.smith.top.model.DocumentPage;
-import care.smith.top.top_document_query.adapter.TextAdapter;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.http.ResponseEntity;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 
-import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 class DocumentApiDelegateImplTest extends AbstractNLPTest {
-  private DocumentApiDelegateImpl documentApi;
+  private static DocumentApiDelegateImpl documentApi;
   Set<Document> documents1 = Set.of(
       new Document().id("d1").name("Document 1"));
   Set<Document> documents2 = Set.of(
@@ -34,8 +26,8 @@ class DocumentApiDelegateImplTest extends AbstractNLPTest {
   Set<Document> documents1_2 = Set.of(
       new Document().id("d1").name("Document 1"), new Document().id("d2").name("Document 2"));
 
-  @BeforeEach
-  void setUp() throws IOException, InstantiationException {
+  @BeforeAll
+  static void setUp() throws IOException, InstantiationException {
     documentApi = new DocumentApiDelegateImpl(mockedDocumentService());
   }
 
@@ -85,54 +77,5 @@ class DocumentApiDelegateImplTest extends AbstractNLPTest {
     ResponseEntity<Document> response2 =  documentApi.getSingleDocumentById(
         "d2", "exampleDataSource", null);
     Assertions.assertEquals(documents2, Set.of(Objects.requireNonNull(response2.getBody())));
-  }
-
-  private static DocumentService mockedDocumentService() throws IOException, InstantiationException {
-    PageImpl<Document> page1 = new PageImpl<>(List.of(new Document().id("d1").name("Document 1")));
-    PageImpl<Document> page2 = new PageImpl<>(List.of(new Document().id("d2").name("Document 2")));
-    PageImpl<Document> page1_2 = new PageImpl<>(
-        List.of(new Document().id("d1").name("Document 1"), new Document().id("d2").name("Document 2")));
-    DocumentNodeEntity d1 = new DocumentNodeEntity("d1", "Document 1", Set.of());
-    DocumentNodeEntity d2 = new DocumentNodeEntity("d2", "Document 2", Set.of());
-
-    TextAdapter adapter = mock(TextAdapter.class);
-    when(adapter.getDocumentById("d1"))
-        .thenReturn(Optional.ofNullable(new Document().id("d1").name("Document 1")));
-    when(adapter.getDocumentById("d2"))
-        .thenReturn(Optional.ofNullable(new Document().id("d2").name("Document 2")));
-    when(adapter.getDocumentsByIds(eq(Set.of("d1", "d2")), anyInt())).thenReturn(page1_2);
-    when(adapter.getDocumentsByIds(eq(Set.of("d1")), anyInt())).thenReturn(page1);
-    when(adapter.getDocumentsByIds(eq(Set.of("d2")), anyInt())).thenReturn(page2);
-    when(adapter.getAllDocuments(anyInt())).thenReturn(page1_2);
-    when(adapter.getDocumentsByName(eq("Document 1"), anyInt())).thenReturn(page1);
-    when(adapter.getDocumentsByName(eq("Document 2"), anyInt())).thenReturn(page2);
-    when(adapter.getDocumentsByName(eq("Document*"), anyInt())).thenReturn(page1_2);
-
-    DocumentNodeRepository documentNodeRepository = mock(DocumentNodeRepository.class);
-    when(documentNodeRepository.getDocumentsForPhraseIds(Set.of("p1", "p2"), false))
-        .thenReturn(List.of(d1, d2));
-    when(documentNodeRepository.getDocumentsForPhraseIds(Set.of("p1", "p2"), true))
-        .thenReturn(List.of(d2));
-    when(documentNodeRepository.getDocumentsForConceptIds(Set.of("c1", "c2"), false))
-        .thenReturn(List.of(d1, d2));
-    when(documentNodeRepository.getDocumentsForConceptIds(Set.of("c1"), false))
-        .thenReturn(List.of(d2));
-    when(documentNodeRepository.getDocumentsForConceptIds(Set.of("c2"), false))
-        .thenReturn(List.of(d1, d2));
-
-    DocumentService documentService = mock(DocumentService.class);
-    when(documentService.getDocumentNodeRepository()).thenReturn(documentNodeRepository);
-    when(documentService.getAdapterFromQuery(anyString(), anyString(), any()))
-        .thenReturn(adapter);
-    when(documentService.getAdapterForDataSource(anyString()))
-        .thenReturn(adapter);
-    when(documentService.getDocumentIdsForQuery(anyString(), anyString(), any()))
-        .thenReturn(List.of("d1", "d2"));
-    when(documentService.getDocumentsForConceptIds(anySet(), anyBoolean())).thenCallRealMethod();
-    when(documentService.getDocumentsForConceptIds(anySet(), anyBoolean(), any())).thenCallRealMethod();
-    when(documentService.getDocumentsForPhraseIds(anySet(), anyBoolean())).thenCallRealMethod();
-    when(documentService.getDocumentsForPhraseTexts(anySet(), anyBoolean())).thenCallRealMethod();
-
-    return documentService;
   }
 }
